2008-07-01 16:28 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/harbour-ce-spec   * harbour/harbour-w32-spec     * define whole contrib library list for new contrib/Makefile     + added header files to respect differences between platform       native and Windows only headers
  * harbour/harbour.spec
    * define whole contrib library list for new contrib/Makefile
    * divided libraries which needs external references into
      separated packages
    ! removed unnecessary dependences in binaries rebuild in shared
      mode

  * harbour/contrib/hbodbc/odbc.c
    * pacified warnings
    % minor optimization

  * harbour/contrib/hbcurl/hbcurl.c
  * harbour/contrib/hbfbird/firebird.c
  * harbour/source/vm/extrap.c
    * casting to pacify warnings

  * harbour/contrib/gtwvg/gtwvg.c
    ! fixed pointer comparison - C uses different precedence for !
      operator then Clipper/xbase

  * harbour/source/rtl/hbinet.c
    ! fixed GPF HB_INET[GS]ET{SND,RCV}BUFSIZE() when wrong
      socket parameter is passed. Please remember that after
      RT error the control is returned to application.
    % removed any hardcoded limits for size of data passed in
      single read/write call during normal stream operations
    - removed iSndBufSize/iRcvBufSize

+++ harbour/contrib/hbcurl/hbcurl.c
@@ -201,7 +201,7 @@ size_t hb_curl_read_file_callback( void * buffer, size_t size, size_t nmemb, voi
          hb_curl->ul_handle = hb_fsOpen( hb_curl->ul_name, FO_READ );
 
          if( hb_curl->ul_handle == FS_ERROR )
-            return -1;
+            return ( size_t ) -1;
       }
 
       ret = ( size_t ) hb_fsReadLarge( hb_curl->ul_handle, ( BYTE * ) buffer, size * nmemb );
@@ -209,7 +209,7 @@ size_t hb_curl_read_file_callback( void * buffer, size_t size, size_t nmemb, voi
       return hb_fsError() ? CURL_READFUNC_ABORT : ret;
    }
 
-   return -1;
+   return ( size_t ) -1;
 }
 
 size_t hb_curl_read_buff_callback( void * buffer, size_t size, size_t nmemb, void * Cargo )
@@ -231,7 +231,7 @@ size_t hb_curl_read_buff_callback( void * buffer, size_t size, size_t nmemb, voi
       return nTodo;
    }
 
-   return -1;
+   return ( size_t ) -1;
 }
 
 size_t hb_curl_write_file_callback( void * buffer, size_t size, size_t nmemb, void * Cargo )
@@ -245,13 +245,13 @@ size_t hb_curl_write_file_callback( void * buffer, size_t size, size_t nmemb, vo
          hb_curl->dl_handle = hb_fsCreate( hb_curl->dl_name, FC_NORMAL );
 
          if( hb_curl->dl_handle == FS_ERROR )
-            return -1;
+            return ( size_t ) -1;
       }
 
       return hb_fsWriteLarge( hb_curl->dl_handle, ( BYTE * ) buffer, size * nmemb );
    }
 
-   return -1;
+   return ( size_t ) -1;
 }
 
 #define HB_CURL_DL_BUFF_SIZE_INIT ( CURL_MAX_WRITE_SIZE * 2 )
@@ -279,7 +279,7 @@ size_t hb_curl_write_buff_callback( void * buffer, size_t size, size_t nmemb, vo
       return nTodo;
    }
 
-   return -1;
+   return ( size_t ) -1;
 }
 
 int hb_curl_progress_callback( void * Cargo, double dltotal, double dlnow, double ultotal, double ulnow )
